RadioMessagingProxy: Fix sending CDMA IMS SMS
When cdmaPdu is not null, we need to set the message technology to
THREE_GPP2 so that dispatchImsCdmaSms() is chosen in libril. If we
don't do this, libril chooses dispatchImsGsmSms() instead, which
means that the message is empty if AIDL is being used and may be
empty or invalid if HIDL is being used.

Specifically, this fixes (some?) devices not being able to send
text messages on Verizon with an "error 0" message.

Fixes: https://gitlab.com/LineageOS/issues/android/-/issues/5391
Fixes: https://gitlab.com/LineageOS/issues/android/-/issues/5612

Signed-off-by: Pranav Temkar <pranavtemkar@gmail.com>
Change-Id: If287c6fa1712a624531e11dd09aeaf7a6c7914c6